FIX https://github.com/OWASP/www-project-secure-headers/pull/297
